+++ cpp/src/arrow/ipc/api.h
@@ -18,9 +18,10 @@
 #ifndef ARROW_IPC_API_H
 #define ARROW_IPC_API_H
 
+#include "arrow/ipc/dictionary.h"
 #include "arrow/ipc/feather.h"
 #include "arrow/ipc/json.h"
-#include "arrow/ipc/metadata.h"
+#include "arrow/ipc/message.h"
 #include "arrow/ipc/reader.h"
 #include "arrow/ipc/writer.h"
