+++ cpp/src/arrow/ipc/message.h
@@ -17,8 +17,8 @@
 
 // C++ object model and user API for interprocess schema messaging
 
-#ifndef ARROW_IPC_METADATA_H
-#define ARROW_IPC_METADATA_H
+#ifndef ARROW_IPC_MESSAGE_H
+#define ARROW_IPC_MESSAGE_H
 
 #include <cstdint>
 #include <memory>
@@ -26,6 +26,7 @@
 #include <unordered_map>
 #include <vector>
 
+#include "arrow/status.h"
 #include "arrow/util/macros.h"
 #include "arrow/util/visibility.h"
 
@@ -33,11 +34,7 @@ namespace arrow {
 
 class Array;
 class Buffer;
-class DataType;
 class Field;
-class Schema;
-class Status;
-class Tensor;
 
 namespace io {
 
@@ -51,96 +48,16 @@ namespace ipc {
 
 enum class MetadataVersion : char { V1, V2, V3 };
 
-static constexpr const char* kArrowMagicBytes = "ARROW1";
-
 // ARROW-109: We set this number arbitrarily to help catch user mistakes. For
 // deeply nested schemas, it is expected the user will indicate explicitly the
 // maximum allowed recursion depth
 constexpr int kMaxNestingDepth = 64;
-
-struct ARROW_EXPORT FieldMetadata {
-  int64_t length;
-  int64_t null_count;
-  int64_t offset;
-};
-
-struct ARROW_EXPORT BufferMetadata {
-  /// The shared memory page id where to find this. Set to -1 if unused
-  int32_t page;
-
-  /// The relative offset into the memory page to the starting byte of the buffer
-  int64_t offset;
-
-  /// Absolute length in bytes of the buffer
-  int64_t length;
-};
-
-struct FileBlock {
-  int64_t offset;
-  int32_t metadata_length;
-  int64_t body_length;
-};
-
-//----------------------------------------------------------------------
-
-using DictionaryMap = std::unordered_map<int64_t, std::shared_ptr<Array>>;
-using DictionaryTypeMap = std::unordered_map<int64_t, std::shared_ptr<Field>>;
-
-// Memoization data structure for handling shared dictionaries
-class ARROW_EXPORT DictionaryMemo {
- public:
-  DictionaryMemo();
-
-  // Returns KeyError if dictionary not found
-  Status GetDictionary(int64_t id, std::shared_ptr<Array>* dictionary) const;
-
-  /// Return id for dictionary, computing new id if necessary
-  int64_t GetId(const std::shared_ptr<Array>& dictionary);
-
-  bool HasDictionary(const std::shared_ptr<Array>& dictionary) const;
-  bool HasDictionaryId(int64_t id) const;
-
-  // Add a dictionary to the memo with a particular id. Returns KeyError if
-  // that dictionary already exists
-  Status AddDictionary(int64_t id, const std::shared_ptr<Array>& dictionary);
-
-  const DictionaryMap& id_to_dictionary() const { return id_to_dictionary_; }
-
-  int size() const { return static_cast<int>(id_to_dictionary_.size()); }
-
- private:
-  // Dictionary memory addresses, to track whether a dictionary has been seen
-  // before
-  std::unordered_map<intptr_t, int64_t> dictionary_to_id_;
-
-  // Map of dictionary id to dictionary array
-  DictionaryMap id_to_dictionary_;
-
-  DISALLOW_COPY_AND_ASSIGN(DictionaryMemo);
-};
 
 // Read interface classes. We do not fully deserialize the flatbuffers so that
 // individual fields metadata can be retrieved from very large schema without
 //
 
 class Message;
-
-// Retrieve a list of all the dictionary ids and types required by the schema for
-// reconstruction. The presumption is that these will be loaded either from
-// the stream or file (or they may already be somewhere else in memory)
-Status GetDictionaryTypes(const void* opaque_schema, DictionaryTypeMap* id_to_field);
-
-// Construct a complete Schema from the message. May be expensive for very
-// large schemas if you are only interested in a few fields
-Status ARROW_EXPORT GetSchema(const void* opaque_schema,
-                              const DictionaryMemo& dictionary_memo,
-                              std::shared_ptr<Schema>* out);
-
-Status ARROW_EXPORT GetTensorMetadata(const Buffer& metadata,
-                                      std::shared_ptr<DataType>* type,
-                                      std::vector<int64_t>* shape,
-                                      std::vector<int64_t>* strides,
-                                      std::vector<std::string>* dim_names);
 
 /// \brief An IPC message including metadata and body
 class ARROW_EXPORT Message {
@@ -219,15 +136,26 @@ class ARROW_EXPORT MessageReader {
 /// \since 0.5.0
 class ARROW_EXPORT InputStreamMessageReader : public MessageReader {
  public:
-  explicit InputStreamMessageReader(const std::shared_ptr<io::InputStream>& stream)
-      : stream_(stream) {}
+  explicit InputStreamMessageReader(io::InputStream* stream) : stream_(stream) {}
+
+#ifndef ARROW_NO_DEPRECATED_API
+  /// \deprecated Since 0.7.0
+  explicit InputStreamMessageReader(const std::shared_ptr<io::InputStream>& owned_stream)
+      : InputStreamMessageReader(owned_stream.get()) {
+    owned_stream_ = owned_stream;
+  }
+#endif
 
   ~InputStreamMessageReader();
 
   Status ReadNextMessage(std::unique_ptr<Message>* message) override;
 
  private:
-  std::shared_ptr<io::InputStream> stream_;
+  io::InputStream* stream_;
+
+#ifndef ARROW_NO_DEPRECATED_API
+  std::shared_ptr<io::InputStream> owned_stream_;
+#endif
 };
 
 /// \brief Read encapulated RPC message from position in file
@@ -253,48 +181,10 @@ Status ReadMessage(const int64_t offset, const int32_t metadata_length,
 /// Read length-prefixed message with as-yet unknown length. Returns nullptr if
 /// there are not enough bytes available or the message length is 0 (e.g. EOS
 /// in a stream)
-Status ARROW_EXPORT ReadMessage(io::InputStream* stream,
-                                std::unique_ptr<Message>* message);
-
-/// Write a serialized message metadata with a length-prefix and padding to an
-/// 8-byte offset
-///
-/// <message_size: int32><message: const void*><padding>
-Status ARROW_EXPORT WriteMessage(const Buffer& message, io::OutputStream* file,
-                                 int32_t* message_length);
-
-// Serialize arrow::Schema as a Flatbuffer
-//
-// \param[in] schema a Schema instance
-// \param[inout] dictionary_memo class for tracking dictionaries and assigning
-// dictionary ids
-// \param[out] out the serialized arrow::Buffer
-// \return Status outcome
-Status ARROW_EXPORT WriteSchemaMessage(const Schema& schema,
-                                       DictionaryMemo* dictionary_memo,
-                                       std::shared_ptr<Buffer>* out);
-
 ARROW_EXPORT
-Status WriteRecordBatchMessage(const int64_t length, const int64_t body_length,
-                               const std::vector<FieldMetadata>& nodes,
-                               const std::vector<BufferMetadata>& buffers,
-                               std::shared_ptr<Buffer>* out);
-
-ARROW_EXPORT
-Status WriteTensorMessage(const Tensor& tensor, const int64_t buffer_start_offset,
-                          std::shared_ptr<Buffer>* out);
-
-Status WriteDictionaryMessage(const int64_t id, const int64_t length,
-                              const int64_t body_length,
-                              const std::vector<FieldMetadata>& nodes,
-                              const std::vector<BufferMetadata>& buffers,
-                              std::shared_ptr<Buffer>* out);
-
-Status WriteFileFooter(const Schema& schema, const std::vector<FileBlock>& dictionaries,
-                       const std::vector<FileBlock>& record_batches,
-                       DictionaryMemo* dictionary_memo, io::OutputStream* out);
+Status ReadMessage(io::InputStream* stream, std::unique_ptr<Message>* message);
 
 }  // namespace ipc
 }  // namespace arrow
 
-#endif  // ARROW_IPC_METADATA_H
+#endif  // ARROW_IPC_MESSAGE_H
