+++ cpp/src/arrow/ipc/reader.h
@@ -24,7 +24,7 @@
 #include <memory>
 #include <vector>
 
-#include "arrow/ipc/metadata.h"
+#include "arrow/ipc/message.h"
 #include "arrow/util/visibility.h"
 
 namespace arrow {
@@ -69,18 +69,28 @@ class ARROW_EXPORT RecordBatchStreamReader : public RecordBatchReader {
   /// Create batch reader from generic MessageReader
   ///
   /// \param(in) message_reader a MessageReader implementation
-  /// \param(out) out the created RecordBatchStreamReader object
+  /// \param(out) out the created RecordBatchReader object
   /// \return Status
+  static Status Open(std::unique_ptr<MessageReader> message_reader,
+                     std::shared_ptr<RecordBatchReader>* out);
+
+#ifndef ARROW_NO_DEPRECATED_API
+  /// \deprecated Since 0.7.0
   static Status Open(std::unique_ptr<MessageReader> message_reader,
                      std::shared_ptr<RecordBatchStreamReader>* out);
 
-  /// \Create Record batch stream reader from InputStream
-  ///
-  /// \param(in) stream an input stream instance
-  /// \param(out) out the created RecordBatchStreamReader object
-  /// \return Status
+  /// \deprecated Since 0.7.0
   static Status Open(const std::shared_ptr<io::InputStream>& stream,
                      std::shared_ptr<RecordBatchStreamReader>* out);
+#endif
+
+  /// \brief Record batch stream reader from InputStream
+  ///
+  /// \param(in) stream an input stream instance. Must stay alive throughout
+  /// lifetime of stream reader
+  /// \param(out) out the created RecordBatchStreamReader object
+  /// \return Status
+  static Status Open(io::InputStream* stream, std::shared_ptr<RecordBatchReader>* out);
 
   std::shared_ptr<Schema> schema() const override;
   Status ReadNextRecordBatch(std::shared_ptr<RecordBatch>* batch) override;
@@ -142,6 +152,27 @@ class ARROW_EXPORT RecordBatchFileReader {
 
 // Generic read functions; does not copy data if the input supports zero copy reads
 
+/// \brief Read Schema from stream serialized as a sequence of IPC messages
+///
+/// \param[in] stream an InputStream
+/// \param[out] out the output Schema
+///
+/// If record batches follow the schema, it is better to use
+/// RecordBatchStreamReader
+ARROW_EXPORT
+Status ReadSchema(io::InputStream* stream, std::shared_ptr<Schema>* out);
+
+/// Read record batch as encapsulated IPC message with metadata size prefix and
+/// header
+///
+/// \param(in) schema the record batch schema
+/// \param(in) offset the file location of the start of the message
+/// \param(in) file the file where the batch is located
+/// \param(out) out the read record batch
+ARROW_EXPORT
+Status ReadRecordBatch(const std::shared_ptr<Schema>& schema, io::InputStream* stream,
+                       std::shared_ptr<RecordBatch>* out);
+
 /// \brief Read record batch from file given metadata and schema
 ///
 /// \param(in) metadata a Message containing the record batch metadata
@@ -174,17 +205,6 @@ Status ReadRecordBatch(const Buffer& metadata, const std::shared_ptr<Schema>& sc
                        int max_recursion_depth, io::RandomAccessFile* file,
                        std::shared_ptr<RecordBatch>* out);
 
-/// Read record batch as encapsulated IPC message with metadata size prefix and
-/// header
-///
-/// \param(in) schema the record batch schema
-/// \param(in) offset the file location of the start of the message
-/// \param(in) file the file where the batch is located
-/// \param(out) out the read record batch
-ARROW_EXPORT
-Status ReadRecordBatch(const std::shared_ptr<Schema>& schema, int64_t offset,
-                       io::RandomAccessFile* file, std::shared_ptr<RecordBatch>* out);
-
 /// EXPERIMENTAL: Read arrow::Tensor as encapsulated IPC message in file
 ///
 /// \param(in) offset the file location of the start of the message
@@ -193,6 +213,15 @@ Status ReadRecordBatch(const std::shared_ptr<Schema>& schema, int64_t offset,
 ARROW_EXPORT
 Status ReadTensor(int64_t offset, io::RandomAccessFile* file,
                   std::shared_ptr<Tensor>* out);
+
+#ifndef ARROW_NO_DEPRECATED_API
+/// \deprecated Since 0.7.0
+///
+/// Deprecated in favor of more general InputStream-based API
+ARROW_EXPORT
+Status ReadRecordBatch(const std::shared_ptr<Schema>& schema, int64_t offset,
+                       io::RandomAccessFile* stream, std::shared_ptr<RecordBatch>* out);
+#endif
 
 }  // namespace ipc
 }  // namespace arrow
